DTS: Description:单位统一
Team:
Feature or Bugfix:

Match-id-ad32f2f4dc34a831761620550ef059824bc25096

+++ test/msprof_cpp/analysis_ut/viewer/test/hbm_processor_utest.cpp
@@ -120,7 +120,7 @@ TEST_F(HBMProcessorUTest, TestRunShouldReturnFalseWhenInsertDataFailed)
 
 TEST_F(HBMProcessorUTest, TestRunShouldReturnFalseWhenReserveFailedThenDataIsEmpty)
 {
-    using TempT = std::tuple<uint16_t, uint64_t, double, uint8_t, uint16_t>;
+    using TempT = std::tuple<uint16_t, uint64_t, uint64_t, uint8_t, uint16_t>;
     MOCKER_CPP(&std::vector<TempT>::reserve).stubs().will(throws(std::bad_alloc()));
     auto processor = HBMProcessor(DB_PATH, PROF_PATHS);
     EXPECT_FALSE(processor.Run());
