+++ test/msprof_cpp/analysis_ut/viewer/test/hccs_processor_utest.cpp
@@ -23,9 +23,9 @@ using namespace Analysis::Utils;
 // device_id, timestamp, txthroughput, rxthroughput
 using HccsDataFormat = std::vector<std::tuple<uint32_t, double, double, double>>;
 // deviceId, timestampNs, txThroughput, rxThroughput
-using ProcessedDataFormat = std::vector<std::tuple<uint16_t, uint64_t, double, double>>;
+using ProcessedDataFormat = std::vector<std::tuple<uint16_t, uint64_t, uint64_t, uint64_t>>;
 // deviceId, timestampNs, txThroughput, rxThroughput
-using QueryDataFormat = std::vector<std::tuple<uint32_t, uint64_t, double, double>>;
+using QueryDataFormat = std::vector<std::tuple<uint32_t, uint64_t, uint64_t, uint64_t>>;
 
 const std::string HCCS_DIR = "./hccs";
 const std::string REPORT = "report.db";
